Renamed remaining module to utils
