+++ tests/unit_tests/providers/opensearch_article_recommendation_test.py
@@ -5,7 +5,7 @@ from sciety_labs.providers.interfaces.article_recommendation import (
     ArticleRecommendationFields,
     ArticleRecommendationFilterParameters
 )
-from sciety_labs.providers.opensearch.opensearch_article_recommendation import (
+from sciety_labs.providers.opensearch.utils import (
     ARTICLE_TITLE_OPENSEARCH_FIELDS,
     AUTHOR_LIST_OPENSEARCH_FIELDS,
     EVALUATION_COUNT_OPENSEARCH_FIELDS,
